[SC-170] Move data3Dtreedelegate to own folder
