+++ libraries/disp3D/engine/delegate/data3Dtreedelegate.h
@@ -44,7 +44,7 @@
 //=============================================================================================================
 
 #include "../../disp3D_global.h"
-#include "items/common/abstracttreeitem.h"
+#include "../model/items/common/abstracttreeitem.h"
 
 
 //*************************************************************************************************************
